Enable forced appearance of aggreagate row
The current implementation of aggregate rows in tables only displayes
thew aggregate row if the group has more than one entry. In the instance
of users displaying tested data using row groups displaying the
aggregate row even for a group of 1 could be useful. This adds an
optional flag that forces the appearance of this row for all groups,
even those of size 1.